+++ scripts/generate-entrypoints.mjs
@@ -10,6 +10,7 @@ const categories = [
   "data",
   "date",
   "http",
+  "hash",
   "input",
   "number",
   "object",