+++ types/hash/crc32.d.ts
@@ -1,0 +1,1 @@
+export { crc32, crc32 as default } from "../hash.js";