+++ types/hash/sha256Hex.d.ts
@@ -1,0 +1,1 @@
+export { sha256Hex, sha256Hex as default } from "../hash.js";