+++ types/hash/sha256Json.d.ts
@@ -1,0 +1,1 @@
+export { sha256Json, sha256Json as default } from "../hash.js";