+++ types/hash/stableJsonId.d.ts
@@ -1,0 +1,1 @@
+export { stableJsonId, stableJsonId as default } from "../hash.js";